requirements.txt
streamlit
opencv-python
ultralytics
Pillow
numpy
pyttsx3
